Update with Gamma, ChangePrecision package

+++ src/Poisson.jl
@@ -15,15 +15,15 @@ using CUDA: i32
 function factorial_lookup_up_to_9(n::FloatType)
     n == zero(FloatType) && return one(FloatType)
     n ==  one(FloatType) && return one(FloatType)
-    n == FloatType(2.0)  && return FloatType(2.0)
-    n == FloatType(3.0)  && return FloatType(6.0)
-    n == FloatType(4.0)  && return FloatType(24.0)
-    n == FloatType(5.0)  && return FloatType(120.0)
-    n == FloatType(6.0)  && return FloatType(720.0)
-    n == FloatType(7.0)  && return FloatType(5040.0)
-    n == FloatType(8.0)  && return FloatType(40320.0)
-    n == FloatType(9.0)  && return FloatType(362880.0)
-    n >= FloatType(10.0) && return NaN
+    n == 2.0  && return 2.0
+    n == 3.0  && return 6.0
+    n == 4.0  && return 24.0
+    n == 5.0  && return 120.0
+    n == 6.0  && return 720.0
+    n == 7.0  && return 5040.0
+    n == 8.0  && return 40320.0
+    n == 9.0  && return 362880.0
+    n >= 10.0 && return NaN
 end
 
 struct PoissonADSampler{T<:Real}
@@ -36,7 +36,7 @@ end
 function PoissonADSampler(μ::Real)
     s = sqrt(μ)
     d = 6 * μ^2
-    L = round( μ - 1.1484f0, RoundDown)
+    L = round( μ - 1.1484, RoundDown)
 
     PoissonADSampler(promote(μ, s, d)..., L)
 end
@@ -65,29 +65,29 @@ function rand(rng::AbstractRNG, sampler::PoissonADSampler)
         end
 
 
-        ω  = FloatType(0.3989422804014327)   / s
-        b1 = FloatType(0.041666666666666664) / μ
-        b2 = FloatType(0.3) * b1 * b1
-        c3 = FloatType(0.14285714285714285) * b1 * b2
+        ω  = 0.3989422804014327   / s
+        b1 = 0.041666666666666664 / μ
+        b2 = 0.3 * b1 * b1
+        c3 = 0.14285714285714285 * b1 * b2
         c2 = b2 - 15 * c3
         c1 = b1 - 6 * b2 + 45 * c3
         c0 = 1 - b1 + 3 * b2 - 15 * c3
 
-        if K < FloatType(10.0)
+        if K < 10.0
             px = -μ
             py =  μ^K/factorial_lookup_up_to_9(K)
         else
-            δ  = FloatType(0.08333333333333333) / K
-            δ -= FloatType(4.8)*δ^3
+            δ  = 0.08333333333333333 / K
+            δ -= 4.8 * δ^3
             V  = (μ-K) / K
             px = K*log1pmx(V) - δ # avoids need for table
-            py = FloatType(0.3989422804014327) / sqrt(K)
+            py = 0.3989422804014327 / sqrt(K)
         end
 
-        X  = (K-μ+FloatType(0.5)) / s
-        X2 =  X^2
-        fx = -X2 / 2 # missing negation in pseudo-algorithm, but appears in fortran code.
-        fy =  ω*(((c3*X2+c2)*X2+c1)*X2+c0)
+        X  = ( K- μ + 0.5) / s
+        X2 =   X^2
+        fx =  -X2 / 2 # missing negation in pseudo-algorithm, but appears in fortran code.
+        fy =   ω* (((c3*X2 + c2)*X2 + c1)*X2 + c0)
 
         # Step Q
         if fy * (1 - U) <= py * exp(px - fx)
@@ -99,39 +99,39 @@ function rand(rng::AbstractRNG, sampler::PoissonADSampler)
         # Step E
         E = randexp(rng, μType)
         U = 2 * Random.rand(rng, μType) - one(μType)
-        T = FloatType(1.8) + copysign(E, U)
-        if T <= FloatType(-0.6744)
+        T = 1.8 + copysign(E, U)
+        if T <= -0.6744
             continue
         end
 
         K = round( μ + s * T, RoundDown)
         #px, py, fx, fy = procf(μ, K, s)
 
-        ω  = FloatType(0.3989422804014327)   / s
-        b1 = FloatType(0.041666666666666664) / μ
-        b2 = FloatType(0.3) * b1 * b1
-        c3 = FloatType(0.14285714285714285) * b1 * b2
+        ω  = 0.3989422804014327   / s
+        b1 = 0.041666666666666664 / μ
+        b2 = 0.3 * b1 * b1
+        c3 = 0.14285714285714285 * b1 * b2
         c2 = b2 - 15 * c3
         c1 = b1 - 6 * b2 + 45 * c3
         c0 = 1 - b1 + 3 * b2 - 15 * c3
 
-        if K < FloatType(10.0)
+        if K < 10.0
             px = -μ
             py =  μ^K/factorial_lookup_up_to_9(K)
         else
-            δ  = FloatType(0.08333333333333333) /K
-            δ -= FloatType(4.8)*δ^3
+            δ  = 0.08333333333333333 /K
+            δ -= 4.8 * δ^3
             V  = (μ-K) / K
             px = K*log1pmx(V) - δ # avoids need for table
-            py = FloatType(0.3989422804014327) / sqrt(K)
+            py = 0.3989422804014327 / sqrt(K)
         end
 
-        X  = (K-μ+FloatType(0.5)) / s
+        X  = (K - μ + 0.5) / s
         X2 =  X^2
         fx = -X2 / 2 # missing negation in pseudo-algorithm, but appears in fortran code.
         fy =  ω*(((c3*X2+c2)*X2+c1)*X2+c0)
 
-        c = FloatType(0.1069) / μ
+        c = 0.1069 / μ
 
         # Step H
         if c*abs(U) <= py*exp(px + E) - fy*exp(fx + E)
